shop/requirements.txt
Django==3.2.13
django-environ==0.4.5
django-psdb-engine==1.0.0
mysqlclient==2.0.3
djangorestframework==3.12.4
python-jose==3.2.0
requests==2.25.1